01.21

+++ src/main/java/com/ch/project/model/Rating.java
@@ -7,5 +7,5 @@ public class Rating {
 	private int 	r_no;			// 평점 번호
 	private String 	m_id; 			// 아이디
 	private int 	b_no;			// 활동한 게시글 번호
-	private int 	r_score;		// 평점
+	private Float 	r_score;		// 평점
 }
